Squelch warnings for negating an unsigned value.

+++ libbpf/bpf_impl.c
@@ -339,7 +339,12 @@ load_byte:
 				continue;
 
 			case BPF_ALU|BPF_NEG:
-				A = -A;
+				/*
+				 * Most BPF arithmetic is unsigned, but negation
+				 * can't be unsigned; throw some casts to
+				 * specify what we're trying to do.
+				 */
+				A = (uint32_t)(-(int32_t)A);
 				continue;
 
 			case BPF_MISC|BPF_TAX:
